Move interface maps to a dedicated module
